Rename: 폴더 이름 소문자로 변경
- Modal -> modal
- BackTesting -> backtesting
