+++ src/components/backtesting/FactorItem.tsx
