+++ src/components/backtesting/FactorSetup.tsx
