+++ src/components/backtesting/TechnicalAnalysisStrategySetup.tsx
